# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Vocabulary API
  version: 0.3.0
  description: BRD-aligned API contract for Dictionary Platform
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/vocabulary:
    get:
      summary: List / search vocabularies (public)
      tags: [Vocabulary]
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Filter by word or translation (case-insensitive)
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Vocabulary list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyListResponse'
    post:
      summary: Add vocabulary (protected)
      tags: [Vocabulary]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVocabularyRequest'
      responses:
        '201':
          description: Vocabulary created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/admin/vocabulary:
    get:
      summary: Admin list/search vocabularies
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: word
          schema:
            type: string
        - in: query
          name: translation
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, blocked]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Admin vocabulary list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Admin create vocabulary
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVocabularyRequest'
      responses:
        '201':
          description: Vocabulary created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/vocabulary/{id}:
    get:
      summary: Admin get vocabulary by id
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vocabulary details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Admin update vocabulary
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVocabularyRequest'
      responses:
        '200':
          description: Vocabulary updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Admin delete vocabulary
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Vocabulary deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/admin/vocabulary/{id}/approve:
    post:
      summary: Admin approve vocabulary
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vocabulary approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/admin/vocabulary/{id}/reject:
    post:
      summary: Admin reject vocabulary
      tags: [AdminVocabulary]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vocabulary rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabularyItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateVocabularyRequest:
      type: object
      required: [word, translation]
      properties:
        word:
          type: string
        translation:
          type: string
        example:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, blocked]
    VocabularyItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        word:
          type: string
        translation:
          type: string
        example:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, blocked]
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    VocabularyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VocabularyItem'
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request body or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid token / credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'